+++ docs/src/Main.tsx
@@ -7,6 +7,7 @@ import {
   Switch as ReactSwitch,
 } from 'react-router-dom';
 import Alert from '../../packages/alert/docs/Alert.mdx';
+import Breadcrumbs from '../../packages/breadcrumbs/docs/Breadcrumbs.mdx';
 import Box from '../../packages/box/docs/Box.mdx';
 import Button from '../../packages/button/docs/Button.mdx';
 import Expandable from '../../packages/expandable/docs/Expandable.mdx';
@@ -14,7 +15,6 @@ import TextArea from '../../packages/textarea/docs/TextArea.mdx';
 import TextField from '../../packages/textfield/docs/TextField.mdx';
 
 /*
-import Breadcrumbs from '../../packages/breadcrumbs/docs/Breadcrumbs.mdx';
 import Pill from '../../packages/pill/docs/Pill.mdx';
 import Card from '../../packages/card/docs/Card.mdx';
 import Combobox from '../../packages/combobox/docs/Combobox.mdx';
@@ -70,6 +70,9 @@ const App = () => {
           <Route path="/box">
             <Box />
           </Route>
+          <Route path="/breadcrumbs">
+            <Breadcrumbs />
+          </Route>
 
           <Route path="/expandable">
             <Expandable />
@@ -79,10 +82,7 @@ const App = () => {
           <Route path="/modal">
             <Modal />
           </Route>
-
-          <Route path="/breadcrumbs">
-            <Breadcrumbs />
-          </Route>
+        
 
           <Route path="/attention">
             <Attention />